Similarity UDF (#535)
* register open as udf

* fix open and test by manual test (evaluate ndarray is ambiguous)

* reformat

* fix test cases for open

* reformat

* fix test case

* reformat

* add dummy feature extractor

* rebase master

* fix rebase master

* working version of similarity fe

* disable error to pass test

* fix orderby issue

* reformat

* move to existing function

* change query syntax and fix test case after syntax change

* lint error fix

* reformat

* fix col alias

* reformat

* simplify duplicate code

* add test for feature extractor and fix open to meet feature extractor requirement

* fix test case

* reformat

* add comment

* fix lint error

* reformat after rebase

+++ eva/udfs/ndarray/open.py
@@ -43,7 +43,7 @@ class Open(AbstractUDF):
                 data = self._data_cache[path_str]
             else:
                 try:
-                    data = cv2.imread(path_str).astype(np.float32)
+                    data = cv2.imread(path_str)
                 except Exception as e:
                     raise ExecutorError(str(e))
